move Symbol into its own sub module
